refactor: 아이콘 경로 수정 및 Dropdown 버튼 스타일 개선
